QCLI Test automation - merged agent feature test into one
