--- e2etests/tests/test_agent_schema_command.rs
@@ -1,58 +1,0 @@
-use q_cli_e2e_tests::q_chat_helper::QChatSession;
-
-#[test]
-#[cfg(feature = "agent")]
-fn test_agent_schema_command() -> Result<(), Box<dyn std::error::Error>> {
-    println!("🔍 Testing /agent schema...");
-    
-    let mut chat = QChatSession::new()?;
-    println!("✅ Q Chat session started");
-    
-    let response = chat.execute_command("/agent schema")?;
-    
-    println!("📝 Agent schema response: {} bytes", response.len());
-    println!("📝 FULL OUTPUT:");
-    println!("{}", response);
-    println!("📝 END OUTPUT");
-    
-    // Collect all failures instead of stopping at first one
-    let mut failures = Vec::new();
-    
-    // Verify JSON schema root level keys
-    if !response.contains("$schema") { failures.push("Missing $schema key"); }
-    if !response.contains("title") { failures.push("Missing title key"); }
-    if !response.contains("description") { failures.push("Missing description key"); }
-    if !response.contains("type") { failures.push("Missing type key"); }
-    if !response.contains("properties") { failures.push("Missing properties key"); }
-
-    //TODO : Due to response is not printing  fully  below keys are not getting detected. check response logic 
-    // if !response.contains("additionalProperties") { failures.push("Missing additionalProperties key"); }
-    // if !response.contains("required") { failures.push("Missing required key"); }
-    // if !response.contains("$defs") { failures.push("Missing $defs key"); }
-    
-    // Verify some key properties exist
-    // if !response.contains("name") { failures.push("Missing name property"); }
-    // if !response.contains("prompt") { failures.push("Missing prompt property"); }
-    // if !response.contains("mcpServers") { failures.push("Missing mcpServers property"); }
-    // if !response.contains("tools") { failures.push("Missing tools property"); }
-    // if !response.contains("toolAliases") { failures.push("Missing toolAliases property"); }
-    // if !response.contains("allowedTools") { failures.push("Missing allowedTools property"); }
-    // if !response.contains("resources") { failures.push("Missing resources property"); }
-    // if !response.contains("hooks") { failures.push("Missing hooks property"); }
-    // if !response.contains("toolsSettings") { failures.push("Missing toolsSettings property"); }
-    // if !response.contains("useLegacyMcpJson") { failures.push("Missing useLegacyMcpJson property"); }
-    
-    // Report all failures at once
-    if !failures.is_empty() {
-        panic!("Test failures: {}", failures.join(", "));
-    }
-    
-    println!("✅ Found all expected JSON schema keys and properties");
-    
-    println!("✅ /agent schema executed successfully with valid JSON schema");
-    
-    chat.quit()?;
-    println!("✅ Test completed successfully");
-    
-    Ok(())
-}
